dev-requirements.txt
# install latest changes in dbt-core
git+https://github.com/dbt-labs/dbt-core.git@1.8.latest#egg=dbt-core&subdirectory=core
git+https://github.com/dbt-labs/dbt-common.git
git+https://github.com/dbt-labs/dbt-adapters.git
git+https://github.com/dbt-labs/dbt-adapters.git#subdirectory=dbt-tests-adapter

# if version 1.x or greater -> pin to major version
# if version 0.x -> pin to minor
black>=24.3
bumpversion~=0.6.0
click~=8.1
ddtrace==2.3.0
flake8~=6.1
flaky~=3.7
freezegun~=1.3
ipdb~=0.13.13
mypy==1.7.1  # patch updates have historically introduced breaking changes
pip-tools~=7.3
pre-commit~=3.5
pre-commit-hooks~=4.5
pytest~=7.4
pytest-csv~=3.0
pytest-dotenv~=0.5.2
pytest-logbook~=1.2
pytest-xdist~=3.5
pytz~=2023.3
tox~=4.11
types-pytz~=2023.3
types-protobuf~=4.24
types-requests~=2.31
twine~=4.0
wheel~=0.42
pyodps